figurines are 'aware' what room they are in if any, saves and loads successfully walls, carpets and figurines configurations, but display of loaded carpets does not function

+++ Point.pde
@@ -1,9 +1,9 @@
 class Point{//store points to check
   int x, y;
   
-  Point(int ix, int iy) {
-    x = ix;
-    y = iy;
+  Point(float ix, float iy) {
+    x = (int)ix;
+    y = (int)iy;
   }
 
   void display(){
